Players can smack back too

+++ Assets/AI/AILogics/Logic_Any_Enemy.cs
@@ -8,8 +8,11 @@ public class Logic_Any_Enemy : AIAction
     public override List<Unit> EvaluateThis(ref Unit MyUnit, ref List<Unit> allUnits)
     {
         Unit u = MyUnit;
-        var allunits = allUnits.FindAll(x => u.IsEnemyUnit != x.IsEnemyUnit);
-        allunits.RemoveRange(1, allunits.Count - 1);
+        var allunits = allUnits.FindAll(x => u.IsEnemyUnit != x.IsEnemyUnit && !x.isDead);
+        if (allunits.Count > 1)
+        {
+            allunits.RemoveRange(1, allunits.Count - 1);
+        }
         return allunits;
     }
